Varios errores arreglados.

+++ SegundoParcial/Entidades/Vehiculos.cs
@@ -12,7 +12,7 @@ namespace SegundoParcial.Entidades
         [Key]
         public int VehiculoId { get; set; }
         public string Descripcion { get; set; }
-        public int Mantenimiento { get; set; }
+        public double Mantenimiento { get; set; }
 
 
         public Vehiculos()
